docker01/python/requirements.txt
pandas
scipy
matplotlib
mysql-connector-python
